Updated dataCleanup.py with drafted tfidf work
Added note from Vilhelm on normalizing number of vectors 
added draft code for converting categories string to array
Threw in draft code for tfidf vectorizer
Question that came up in discussion from Johanna - do we perform two separate tfidf processes on the train and test data or in one and then split between train and test? We could end up with some data leakage if we perform tfidf process but could end up with skewed results on the test set if we do them separately.